+++ realm/realm-library/build.gradle
@@ -46,7 +46,7 @@ repositories {
 
 dependencies {
     provided 'io.reactivex:rxjava:1.1.0'
-    compile "com.android.support:recyclerview-v7:23.1.1"
+    compile 'com.android.support:recyclerview-v7:23.1.1'
     compile "io.realm:realm-annotations:${version}"
 
     androidTestCompile 'io.reactivex:rxjava:1.1.0'
