+++ realm/realm-library/src/androidTest/java/io/realm/entities/UnsupportedCollection.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2016 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.realm.entities;
 
 import android.support.annotation.NonNull;
@@ -14,6 +30,10 @@ import io.realm.RealmQuery;
 import io.realm.RealmResults;
 import io.realm.Sort;
 
+/**
+ * Class that implements {@link OrderedRealmCollection} interface and can be used in testing supported
+ * collections along with {@link RealmResults} and {@link io.realm.RealmList}.
+ */
 public class UnsupportedCollection<E extends RealmObject> implements OrderedRealmCollection<E> {
 
     @Override
