+++ realm/realm-library/src/main/java/io/realm/RealmBaseRecyclerAdapter.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2016 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.realm;
 
 import android.content.Context;
@@ -7,23 +23,27 @@ import android.view.LayoutInflater;
 /**
  * The RealmBaseRecyclerAdapter class is an abstract utility class for binding RecyclerView UI elements to Realm data.
  * <p>
- * This adapter will automatically handle any updates to its data and call {@link #notifyDataSetChanged()} as
- * appropriate.
+ * This adapter will automatically handle any updates to its data and call
+ * {@link android.support.v7.widget.RecyclerView.Adapter#notifyDataSetChanged()} as appropriate.
+ * Currently there is no support for RecyclerView's data callback methods like
+ * {@link android.support.v7.widget.RecyclerView.Adapter#notifyItemInserted(int)},
+ * {@link android.support.v7.widget.RecyclerView.Adapter#notifyItemRemoved(int)},
+ * {@link android.support.v7.widget.RecyclerView.Adapter#notifyItemChanged(int)} etc.
+ * It means that, there is no possibility to use default data animations.
  * <p>
  * The RealmAdapter will stop receiving updates if the Realm instance providing the {@link OrderedRealmCollection} is
  * closed.
  *
  * @param <T> type of {@link RealmObject} stored in the adapter.
- * @param <VH> type of {@link RecyclerView.ViewHolder} used in the adapter.
+ * @param <VH> type of {@link android.support.v7.widget.RecyclerView.ViewHolder} used in the adapter.
  */
 public abstract class RealmBaseRecyclerAdapter<T extends RealmObject, VH extends RecyclerView.ViewHolder> extends RecyclerView.Adapter<VH> {
 
-    protected LayoutInflater inflater;
-    protected Context context;
+    protected final LayoutInflater inflater;
+    protected final Context context;
+    private final boolean hasAutoUpdates;
+    private final RealmChangeListener listener;
     private OrderedRealmCollection<T> adapterData;
-    private RealmChangeListener listener;
-
-    private boolean hasAutoUpdates;
 
     public RealmBaseRecyclerAdapter(Context context, OrderedRealmCollection<T> data, boolean autoUpdate) {
         if (context == null) {
@@ -34,14 +54,12 @@ public abstract class RealmBaseRecyclerAdapter<T extends RealmObject, VH extends
         this.adapterData = data;
         this.inflater = LayoutInflater.from(context);
         this.hasAutoUpdates = autoUpdate;
-        if (autoUpdate) {
-           listener = new RealmChangeListener() {
-               @Override
-               public void onChange() {
-                   notifyDataSetChanged();
-               }
-           };
-        }
+        this.listener = hasAutoUpdates ? new RealmChangeListener() {
+            @Override
+            public void onChange() {
+                notifyDataSetChanged();
+            }
+        } : null;
     }
 
     @Override
@@ -62,7 +80,8 @@ public abstract class RealmBaseRecyclerAdapter<T extends RealmObject, VH extends
 
     /**
      * Returns the current ID for an item. Note that item IDs are not stable so you cannot rely on the item ID being the
-     * same after {@link #notifyDataSetChanged()} or {@link #updateData(OrderedRealmCollection)} has been called.
+     * same after {@link android.support.v7.widget.RecyclerView.Adapter#notifyDataSetChanged()}
+     * or {@link #updateData(OrderedRealmCollection)} has been called.
      *
      * @param index position of item in the adapter.
      * @return current item ID.
@@ -79,49 +98,13 @@ public abstract class RealmBaseRecyclerAdapter<T extends RealmObject, VH extends
 
     /**
      * Returns the item associated with the specified position.
+     * Can return null if provided Realm instance by {@link OrderedRealmCollection} is closed.
      *
      * @param index index of the item.
-     * @return the item at the specified position.
+     * @return the item at the specified position, null if adapter data is not valid.
      */
     public T getItem(int index) {
         return isDataValid() ? adapterData.get(index) : null;
-    }
-
-    /**
-     * Removes the item associated with the specified position. This also deletes the objects from the underlying Realm.
-     *
-     * @param index index of the item to remove.
-     */
-    public void removeItem(int index) {
-        if (isDataValid()) {
-            getRealm().beginTransaction();
-            adapterData.deleteFromRealm(index);
-            getRealm().commitTransaction();
-
-            // There is no need to call notify methods if we have listener
-            if (!hasAutoUpdates) {
-                notifyItemRemoved(index);
-            }
-        }
-    }
-
-    /**
-     * Clears all data from the adapter. This also deletes the objects from the underlying Realm.
-     */
-    public void clear() {
-        if (isDataValid()) {
-            getRealm().beginTransaction();
-            int dataSize = adapterData.size();
-            for (int i = dataSize - 1; i >= 0; i--) {
-                adapterData.deleteFromRealm(i);
-            }
-            getRealm().commitTransaction();
-
-            // There is no need to call notify methods if we have listener
-            if (!hasAutoUpdates) {
-                notifyDataSetChanged();
-            }
-        }
     }
 
     /**
@@ -153,43 +136,33 @@ public abstract class RealmBaseRecyclerAdapter<T extends RealmObject, VH extends
         notifyDataSetChanged();
     }
 
-    private boolean isDataValid() {
-        return adapterData != null && adapterData.isValid();
-    }
-
-    private BaseRealm getRealm() {
-        if (adapterData instanceof RealmResults) {
-            RealmResults realmResults = (RealmResults) adapterData;
-            return realmResults.realm;
-        } else if (adapterData instanceof RealmList) {
-            RealmList realmList = (RealmList) adapterData;
-            return realmList.realm;
-        } else {
-            throw new IllegalArgumentException("RealmCollection not supported: " + adapterData.getClass());
-        }
-    }
-
     private void addListener(OrderedRealmCollection<T> data) {
-        if (data instanceof RealmResults) {
-            RealmResults realmResults = (RealmResults) data;
-            realmResults.realm.handlerController.addChangeListenerAsWeakReference(listener);
-        } else if (data instanceof RealmList) {
-            RealmList realmList = (RealmList) data;
-            realmList.realm.handlerController.addChangeListenerAsWeakReference(listener);
-        } else {
-            throw new IllegalArgumentException("RealmCollection not supported: " + data.getClass());
+        BaseRealm realm = getRealm(data);
+        if (realm != null && !realm.isClosed()) {
+            realm.addChangeListener(listener);
         }
     }
 
     private void removeListener(OrderedRealmCollection<T> data) {
+        BaseRealm realm = getRealm(data);
+        if (realm != null && !realm.isClosed()) {
+            realm.removeChangeListener(listener);
+        }
+    }
+
+    private BaseRealm getRealm(OrderedRealmCollection<T> data) {
         if (data instanceof RealmResults) {
             RealmResults realmResults = (RealmResults) data;
-            realmResults.realm.handlerController.removeWeakChangeListener(listener);
+            return realmResults.realm;
         } else if (data instanceof RealmList) {
             RealmList realmList = (RealmList) data;
-            realmList.realm.handlerController.removeWeakChangeListener(listener);
+            return realmList.realm;
         } else {
             throw new IllegalArgumentException("RealmCollection not supported: " + data.getClass());
         }
+    }
+
+    private boolean isDataValid() {
+        return adapterData != null && adapterData.isValid();
     }
 }
